feat: Refactor enums to use TypeScript string literal types for improved type safety and remove deprecated runtime enums and move into enums/ directory in separate files
